Update gpt.py
Updating the forward function in Transformer block.

The code is simple to example the pull request, but still trying my best to explain below:

As per paper: In 'Add & Norm' block of Transformer, Layer Norm is applied on top of input/ residual & output of Self-attention. While in the current code, first layer Norm is applied & then added back to the input/ residual.